add new files and use generator
