+++ skyplane/api/transfer_job.py
@@ -25,6 +25,7 @@ from skyplane.utils import logger
 from skyplane.utils.definitions import MB
 from skyplane.utils.fn import do_parallel
 from skyplane.utils.path import parse_path
+from skyplane.utils.generator import batch_generator
 
 if TYPE_CHECKING:
     from skyplane.api.dataplane import Dataplane
@@ -281,66 +282,6 @@ class Chunker:
             # drain multipart chunk queue and yield with updated chunk IDs
             while not multipart_chunk_queue.empty():
                 yield multipart_chunk_queue.get()
-
-    @staticmethod
-    def batch_generator(gen_in: Generator[T, None, None], batch_size: int) -> Generator[List[T], None, None]:
-        """Batches generator, while handling StopIteration
-
-        :param gen_in: generator that generates chunk requests
-        :type gen_in: Generator
-        """
-        batch = []
-        for item in gen_in:
-            batch.append(item)
-            if len(batch) == batch_size:
-                yield batch
-                batch = []
-        if len(batch) > 0:
-            yield batch
-
-    @staticmethod
-    def prefetch_generator(gen_in: Generator[T, None, None], buffer_size: int) -> Generator[T, None, None]:
-        """
-        Prefetches from generator while handing StopIteration to ensure items yield immediately.
-        Start a thread to prefetch items from the generator and put them in a queue. Upon StopIteration,
-        the thread will add a sentinel value to the queue.
-
-        :param gen_in: generator that generates chunk requests
-        :type gen_in: Generator
-        :param buffer_size: maximum size of the buffer to temporarily store the generators
-        :type buffer_size: int
-        """
-        sentinel = object()
-        queue = Queue(maxsize=buffer_size)
-
-        def prefetch():
-            for item in gen_in:
-                queue.put(item)
-            queue.put(sentinel)
-
-        thread = threading.Thread(target=prefetch, daemon=True)
-        thread.start()
-
-        while True:
-            item = queue.get()
-            if item is sentinel:
-                break
-            yield item
-
-    @staticmethod
-    def tail_generator(gen_in: Generator[T, None, None], out_list: List[T]) -> Generator[T, None, None]:
-        """Tails generator while handling StopIteration
-
-        :param gen_in: generator that generates chunk requests
-        :type gen_in: Generator
-        :param out_list: list of tail generators
-        :type out_list: List
-        """
-        for item in gen_in:
-            out_list.append(item)
-            yield item
-
-
 @dataclass
 class TransferJob(ABC):
     """
